changefeedccl: fix cloud storage sink ordering issue
Background: the cloud storage sink writes files to
cloud storage with names whose first component is
a timestamp. Beyond the normal changefeed ordering
guarantees, there is an additional guarantee that
if you read these files in lexicographic order,
the observed events will respect the changefeed
ordering guarantee.

However, under certain conditions (partially
lagging + restarts), we could emit 2 files with
the same timestamp name component that, if read in
order, resulted in a violation of the changefeed
ordering guarantee.

This patch fixes the issue by adding another
timestamp to the front of the tiebreaker name
component that tracks the oldest MVCC timestamp
for the file. It also adds a regression test for
this specific issue.

Epic: None
Release note: None